Convert from int to size_t.

+++ fc-solve/source/instance.h
@@ -509,7 +509,7 @@ enum
 typedef struct
 {
     fc_solve_solve_for_state_move_func_t *tests;
-    int num_tests;
+    size_t num_tests;
     fcs_tests_group_type_t shuffling_type;
     fc_solve_state_weighting_t weighting;
 } fcs_tests_list_t;
